Added a basic vampire entity class to handle sundamage, villager conversion etc. Also added a system to sync custom entity data. Not sure if it is very efficient, but it should do what is required
