+++ src/main/java/de/teamlapen/vampirism/util/Helper.java
@@ -6,8 +6,6 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 
-import cpw.mods.fml.relauncher.ReflectionHelper;
-import de.teamlapen.vampirism.entity.EntityVampireHunter;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.EntityCreature;
 import net.minecraft.entity.EntityList;
@@ -18,6 +16,12 @@ import net.minecraft.util.MovingObjectPosition;
 import net.minecraft.util.Vec3;
 import net.minecraft.village.Village;
 import net.minecraft.world.World;
+import cpw.mods.fml.common.network.NetworkRegistry.TargetPoint;
+import cpw.mods.fml.common.network.simpleimpl.IMessage;
+import cpw.mods.fml.relauncher.ReflectionHelper;
+import cpw.mods.fml.relauncher.Side;
+import cpw.mods.fml.relauncher.SideOnly;
+import de.teamlapen.vampirism.VampirismMod;
 
 public class Helper {
 	public static class Reflection {
@@ -129,5 +133,10 @@ public class Helper {
 			}
 		}
 		return list;
+	}
+	
+	@SideOnly(Side.SERVER)
+	public static void sendPacketToPlayersAround(IMessage message,Entity e){
+		VampirismMod.modChannel.sendToAllAround(message, new TargetPoint(e.dimension,e.posX,e.posY,e.posZ,100));
 	}
 }
